# rarbg_local/tests/snapshots/test_integration/test_openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media",
    "version": "0.1.0-dev"
  },
  "servers": [
    {
      "url": "{protocol}://localhost:5000/",
      "description": "Development",
      "variables": {
        "protocol": {
          "enum": [
            "http",
            "https"
          ],
          "default": "https"
        }
      }
    },
    {
      "url": "https://media-staging.herokuapps.com/",
      "description": "Staging"
    },
    {
      "url": "https://media.mause.me/",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/delete/{type}/{id}": {
      "get": {
        "summary": "Delete",
        "operationId": "delete",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MediaType"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Api Delete  Type   Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stream/{type}/{tmdb_id}": {
      "get": {
        "summary": "Stream",
        "operationId": "stream",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "series",
                "movie"
              ],
              "type": "string",
              "title": "Type"
            }
          },
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProviderSource"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "name": "episode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Episode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ITorrent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/select/{tmdb_id}/season/{season}/download_all": {
      "get": {
        "summary": "Select",
        "operationId": "select",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadAllResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/download": {
      "post": {
        "summary": "Download Post",
        "operationId": "download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/DownloadPost"
                },
                "type": "array",
                "title": "Things"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/MovieDetailsSchema"
                      },
                      {
                        "$ref": "#/components/schemas/EpisodeDetailsSchema"
                      }
                    ]
                  },
                  "type": "array",
                  "title": "Response Download Post Api Download Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/index": {
      "get": {
        "summary": "Index",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Stats",
        "operationId": "stats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StatsResponse"
                  },
                  "type": "array",
                  "title": "Response Stats Api Stats Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/movie/{tmdb_id}": {
      "get": {
        "summary": "Movie",
        "operationId": "movie",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/torrents": {
      "get": {
        "summary": "Torrents",
        "operationId": "torrents",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/InnerTorrent"
                  },
                  "type": "object",
                  "title": "Response Torrents Api Torrents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/search": {
      "get": {
        "summary": "Search",
        "operationId": "search",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResponse"
                  },
                  "title": "Response Search Api Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers": {
      "get": {
        "summary": "Get Providers",
        "operationId": "get_providers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderSource"
                  },
                  "type": "array",
                  "title": "Response Get Providers Api Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/discover": {
      "get": {
        "summary": "Discover",
        "operationId": "discover",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discover"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/tmdb/configuration": {
      "get": {
        "summary": "Tmdb Configuration",
        "operationId": "tmdb_configuration",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuration"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/tv/{tmdb_id}": {
      "get": {
        "tags": [
          "tv"
        ],
        "summary": "Api Tv",
        "operationId": "api_tv",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tv/{tmdb_id}/season/{season}": {
      "get": {
        "tags": [
          "tv"
        ],
        "summary": "Api Tv Season",
        "operationId": "api_tv_season",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvSeasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitor": {
      "get": {
        "tags": [
          "monitor"
        ],
        "summary": "Monitor Get",
        "operationId": "monitor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MonitorGet"
                  },
                  "type": "array",
                  "title": "Response Monitor Get Api Monitor Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "monitor"
        ],
        "summary": "Monitor Post",
        "operationId": "monitor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/monitor/{monitor_id}": {
      "delete": {
        "tags": [
          "monitor"
        ],
        "summary": "Monitor Delete",
        "operationId": "monitor_delete",
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "monitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Monitor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Monitor Delete Api Monitor  Monitor Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/diagnostics": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "Diagnostics",
        "operationId": "diagnostics",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsRoot"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenIdConnect": [
              "openid"
            ]
          },
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaType": {
        "type": "string",
        "enum": [
          "series",
          "movie"
        ],
        "title": "MediaType"
      },
...
      "MovieDetailsSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "download": {
            "$ref": "#/components/schemas/DownloadSchema"
          }
        },
        "type": "object",
        "required": [
          "id",
          "download"
        ],
        "title": "MovieDetailsSchema"
      },
...
      "ProviderSource": {
        "type": "string",
        "enum": [
          "kickass",
          "horriblesubs",
          "rarbg",
          "torrentscsv",
          "nyaasi",
          "piratebay"
        ],
        "title": "ProviderSource"
      },
...
    },
    "securitySchemes": {
      "OpenIdConnect": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://mause.au.auth0.com//.well-known/openid-configuration"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}